requirements.txt
django
channels
daphne
redis
channels-redis
celery[redis]
django-celery-results
django-celery-beat
djangorestframework
django-rest-framework-gis
drf-spectacular
requests
pandas
geopandas
gtfs-realtime-bindings
djangorestframework
drf-spectacular
python-decouple
psycopg2-binary
gunicorn
